refactor(orchestrator): extract execution governance workflows
