+++ services/local-service/internal/orchestrator/governance.go
@@ -1,0 +1,183 @@
+package orchestrator
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/audit"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/execution"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/runengine"
+)
+
+// buildApprovalRequest creates the normalized approval_request payload. The
+// object must already be protocol-facing here because it is persisted, replayed
+// to transports, and later echoed back through agent.security.respond.
+func buildApprovalRequest(taskID string, taskIntent map[string]any, assessment execution.GovernanceAssessment) map[string]any {
+	arguments := mapValue(taskIntent, "arguments")
+	targetObject := firstNonEmptyString(assessment.TargetObject, stringValue(arguments, "target_path", "workspace_document"))
+	if targetObject == "" {
+		targetObject = "workspace_document"
+	}
+
+	return map[string]any{
+		"approval_id":    fmt.Sprintf("appr_%s_%d", taskID, time.Now().UnixNano()),
+		"task_id":        taskID,
+		"operation_name": firstNonEmptyString(assessment.OperationName, firstNonEmptyString(stringValue(taskIntent, "name", ""), "write_file")),
+		"risk_level":     firstNonEmptyString(assessment.RiskLevel, "red"),
+		"target_object":  targetObject,
+		"reason":         firstNonEmptyString(assessment.Reason, "policy_requires_authorization"),
+		"status":         "pending",
+		"created_at":     time.Now().Format(dateTimeLayout),
+	}
+}
+
+func (s *Service) applyGovernanceAssessment(plan map[string]any, assessment execution.GovernanceAssessment) map[string]any {
+	updatedPlan := cloneMap(plan)
+	if updatedPlan == nil {
+		updatedPlan = map[string]any{}
+	}
+	if len(assessment.ImpactScope) > 0 {
+		updatedPlan["impact_scope"] = cloneMap(assessment.ImpactScope)
+	}
+	if assessment.OperationName != "" {
+		updatedPlan["operation_name"] = assessment.OperationName
+	}
+	if assessment.TargetObject != "" {
+		updatedPlan["target_object"] = assessment.TargetObject
+	}
+	if assessment.RiskLevel != "" {
+		updatedPlan["risk_level"] = assessment.RiskLevel
+	}
+	if assessment.Reason != "" {
+		updatedPlan["risk_reason"] = assessment.Reason
+	}
+	updatedPlan["checkpoint_required"] = assessment.CheckpointRequired
+	return updatedPlan
+}
+
+func (s *Service) assessTaskGovernance(task runengine.TaskRecord, taskIntent map[string]any) (execution.GovernanceAssessment, bool, error) {
+	if s.executor == nil {
+		return execution.GovernanceAssessment{}, false, nil
+	}
+	resultTitle, _, _ := resultSpecFromIntent(taskIntent)
+	return s.executor.AssessGovernance(context.Background(), execution.Request{
+		TaskID:       task.TaskID,
+		RunID:        task.RunID,
+		SourceType:   task.SourceType,
+		Title:        task.Title,
+		Intent:       taskIntent,
+		Snapshot:     snapshotFromTask(task),
+		DeliveryType: resolveTaskDeliveryType(task, taskIntent),
+		ResultTitle:  resultTitle,
+	})
+}
+
+func (s *Service) handleTaskGovernanceDecision(task runengine.TaskRecord, taskIntent map[string]any) (runengine.TaskRecord, map[string]any, bool, error) {
+	assessment, ok, err := s.assessTaskGovernance(task, taskIntent)
+	if err != nil {
+		return task, nil, false, err
+	}
+	if !ok {
+		assessment, ok = s.fallbackGovernanceAssessment(task, taskIntent)
+		if !ok {
+			return task, nil, false, nil
+		}
+	}
+	if assessment.Deny {
+		response, blockedTask, blockErr := s.blockTaskByAssessment(task, assessment)
+		return blockedTask, response, true, blockErr
+	}
+	if !assessment.ApprovalRequired {
+		return task, nil, false, nil
+	}
+	pendingExecution := s.applyGovernanceAssessment(s.buildPendingExecution(task, taskIntent), assessment)
+	approvalRequest := buildApprovalRequest(task.TaskID, taskIntent, assessment)
+	bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", "检测到待授权操作，请先确认。", task.UpdatedAt.Format(dateTimeLayout))
+	updatedTask := runengine.TaskRecord{}
+	changed := false
+	if s.isPreparedRestartAttempt(task) {
+		updatedTask, changed = s.runEngine.MarkPreparedTaskWaitingApprovalWithPlan(task, approvalRequest, pendingExecution, bubble)
+	} else {
+		updatedTask, changed = s.runEngine.MarkWaitingApprovalWithPlan(task.TaskID, approvalRequest, pendingExecution, bubble)
+	}
+	if !changed {
+		return task, nil, false, ErrTaskNotFound
+	}
+	if err := s.persistApprovalRequestState(updatedTask.TaskID, approvalRequest, assessment.ImpactScope); err != nil {
+		return task, nil, false, err
+	}
+	return updatedTask, map[string]any{
+		"task":            taskMap(updatedTask),
+		"bubble_message":  bubble,
+		"delivery_result": nil,
+	}, true, nil
+}
+
+func (s *Service) fallbackGovernanceAssessment(task runengine.TaskRecord, taskIntent map[string]any) (execution.GovernanceAssessment, bool) {
+	if stringValue(taskIntent, "name", "") != "write_file" && !boolValue(mapValue(taskIntent, "arguments"), "require_authorization", false) {
+		return execution.GovernanceAssessment{}, false
+	}
+	plan := s.buildPendingExecution(task, taskIntent)
+	impactScope := s.buildImpactScope(task, plan)
+	return execution.GovernanceAssessment{
+		OperationName:    firstNonEmptyString(stringValue(taskIntent, "name", ""), "write_file"),
+		TargetObject:     impactScopeTarget(impactScope, targetPathFromIntent(taskIntent)),
+		RiskLevel:        "red",
+		ApprovalRequired: true,
+		Reason:           "policy_requires_authorization",
+		ImpactScope:      impactScope,
+	}, true
+}
+
+func (s *Service) blockTaskByAssessment(task runengine.TaskRecord, assessment execution.GovernanceAssessment) (map[string]any, runengine.TaskRecord, error) {
+	bubbleText := governanceInterceptionBubble(assessment)
+	bubble := s.delivery.BuildBubbleMessage(task.TaskID, "status", bubbleText, task.UpdatedAt.Format(dateTimeLayout))
+	updatedTask := runengine.TaskRecord{}
+	ok := false
+	if s.isPreparedRestartAttempt(task) {
+		updatedTask, ok = s.runEngine.BlockPreparedTaskByPolicy(task, assessment.RiskLevel, bubbleText, assessment.ImpactScope, bubble)
+	} else {
+		updatedTask, ok = s.runEngine.BlockTaskByPolicy(task.TaskID, assessment.RiskLevel, bubbleText, assessment.ImpactScope, bubble)
+	}
+	if !ok {
+		return nil, task, ErrTaskNotFound
+	}
+	auditRecord := s.writeGovernanceAuditRecord(updatedTask.TaskID, updatedTask.RunID, "risk", "intercept_operation", bubbleText, impactScopeTarget(assessment.ImpactScope, assessment.TargetObject), "denied")
+	updatedTask = s.appendAuditData(updatedTask, compactAuditRecords(auditRecord), nil)
+	return map[string]any{
+		"task":            taskMap(updatedTask),
+		"bubble_message":  bubble,
+		"delivery_result": nil,
+		"impact_scope":    cloneMap(assessment.ImpactScope),
+	}, updatedTask, nil
+}
+
+func (s *Service) writeGovernanceAuditRecord(taskID, runID, auditType, action, summary, target, result string) map[string]any {
+	if s.audit == nil {
+		return nil
+	}
+	if record, err := s.audit.Write(context.Background(), audit.RecordInput{
+		TaskID:  taskID,
+		RunID:   runID,
+		Type:    auditType,
+		Action:  action,
+		Summary: summary,
+		Target:  target,
+		Result:  result,
+	}); err == nil {
+		return record.Map()
+	}
+	if record, err := s.audit.BuildRecord(audit.RecordInput{
+		TaskID:  taskID,
+		RunID:   runID,
+		Type:    auditType,
+		Action:  action,
+		Summary: summary,
+		Target:  target,
+		Result:  result,
+	}); err == nil {
+		return record.Map()
+	}
+	return nil
+}
